requirements.txt
dvc==3.50.0
faiss-cpu==1.8.0
langchain==0.1.16
langchain-community==0.0.34
langchain-text-splitters==0.0.1
langchainhub==0.1.15
notebook==7.1.3
pandas==2.2.2
pypdf==4.2.0
ragas==0.1.7
rank_bm25==0.2.2
ruamel.yaml==0.18.6
sentence-transformers==2.7.0
streamlit==1.33.0
streamlit_chat==0.1.1
torch==2.1.2
unstructured[pdf]==0.13.3
